Moved parameter setting in Archimedean to downstream classes
Also labelled todo for both Clayton and Gumbel. Parameter setting
moved downstream as each class of copula has different checks
