+++ copulae/archimedean/__init__.py
@@ -1,3 +1,4 @@
 from .clayton import ClaytonCopula
+from .gumbel import GumbelCopula
 
-__all__ = ['ClaytonCopula']
+__all__ = ['ClaytonCopula', 'GumbelCopula']
