+++ copulae/archimedean/abstract.py
@@ -64,21 +64,6 @@ class AbstractArchimedeanCopula(BaseCopula, ABC):
         """
         raise NotImplementedError
 
-    @property
-    def params(self):
-        return self._theta
-
-    @params.setter
-    def params(self, theta: float):
-        theta = float(theta)
-
-        if self.dim == 2 and theta < -1:
-            raise ValueError('theta must be greater than -1 in 2 dimensional clayton copulae')
-        elif self.dim > 2 and theta < 0:
-            raise ValueError('theta must be positive when dim > 2')
-
-        self._theta = theta
-
     @abstractmethod
     def psi(self, s: Array):
         """
